refactor and reuse find_deployments_for_model
